# huff0: Use buffers for 1X decodes